requirements.txt
discord.py[voice]
python-dotenv
youtube_dl